backend/requirements.txt
flasgger==0.9.7.1
APScheduler==3.10.4
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
Flask-Limiter==3.12
dnspython==2.7.0
razorpay==2.0.0
firebase-admin==7.1.0
marshmallow==4.0.0
bleach==6.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
Markdown==3.9
boto3==1.40.54
itsdangerous==2.2.0
eventlet==0.40.3
SQLAlchemy==2.0.43
MarkupSafe==3.0.2
Werkzeug==3.1.3
PyJWT==2.10.1
python-dotenv==1.1.1
fpdf2==2.8.3
